sched/tune: only force actual top app to big cores
If we force top-app's boost value to be >0 we get excellent
jitter values due to top-app being rendered on big cores.

However along with the actual top app a log of crap ends up on
big cluster too. We can avoid this by checking if adj_score is 0
which means that this app is actually the one that's currently
on the screen.

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>